requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
defusedxml==0.7.0rc1
Django==3.1.3
django-allauth==0.43.0
django-bootstrap4==2.3.1
django-location-field==2.1.0
django-phonenumber-field==5.0.0
idna==2.10
oauthlib==3.1.0
phonenumbers==8.12.13
Pillow==8.0.1
PyJWT==2.0.0a1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
